Remove "I decline" button from cookie modal
Since this button only causes an error message to be displayed because users have to accept all the required cookies, it might be a good idea to remove it entirely. This way, it is obvious that you have to accept the cookies and cannot decline them without leaving the website.